chore(webpack-extraction-plugin): move vitest stubs to src/common/
Replace the .cjs registration + relative-path stubs with explicit
Module._resolveFilename redirects to src/common/webpackLoader.cjs and
src/common/constants.cjs. Exclude src/common/ from the production build
via tsconfig.lib.json.

The previous approach registered .cjs in Node's default extensions and
relied on the resolver finding sibling stubs next to the .ts sources;
this version is more targeted (only the two specific request strings
get rerouted) and keeps test-only files out of src/.
