+++ packages/webpack-extraction-plugin/tsconfig.lib.json
@@ -6,6 +6,6 @@
     "declaration": true,
     "types": ["node", "environment"]
   },
-  "exclude": ["__fixtures__/**/*/*.ts", "**/*.spec.ts", "**/*.test.ts", "jest.config.ts"],
+  "exclude": ["__fixtures__/**/*/*.ts", "**/*.spec.ts", "**/*.test.ts", "src/common/**", "jest.config.ts"],
   "include": ["**/*.ts", "virtual-loader/index.js"]
 }
